RAG-SEC/requirements.txt
openai
langchain-community
langchain-core
pydantic
PyPDF2
python-dotenv
faiss-cpu
sentence-transformers
